Rename global components
