+++ src/components/global/footer.vue
